test/smoke: separate generic matrix code
- In prep for peeling off matrix defs as a separate package.
