+++ test/smoke/testmain_test.go
@@ -10,20 +10,6 @@ import (
 
 var pfs *parallelTestFixtureSet
 
-// Matrix returns the defined sous smoke test matrix.
-func Matrix() matrixDef {
-	m := newMatrix()
-	m.AddDimension("store", "GDM storage to use", map[string]interface{}{
-		"db":  true,
-		"git": false,
-	})
-	m.AddDimension("project", "type of project to build", map[string]interface{}{
-		"simple": projects.SingleDockerfile,
-		"split":  projects.SplitBuild,
-	})
-	return m
-}
-
 func TestMain(m *testing.M) {
 	flag.BoolVar(&flags.printMatrix, "ls", false, "list test matrix names")
 	flag.BoolVar(&flags.printDimensions, "dimensions", false, "list test matrix dimensions")
@@ -32,7 +18,7 @@ func TestMain(m *testing.M) {
 	runRealTests := !(flags.printMatrix || flags.printDimensions)
 
 	if flags.printDimensions {
-		Matrix().PrintDimensions()
+		matrix().PrintDimensions()
 	}
 
 	if runRealTests {
